Create SampleInvoicesUnitTest.cs
This is one of my Unit Tests that I created to test and find errors in the QBXML.  The QbProcessor is my own personal Quickbooks requests processor which is sending the xml.  You will need to use your own requests processor for that.  However; the resulting string is then processed through the included ToView method to copy the xml response into a complete C# object.  You can then access the QBXML response as you would any other C# object.